refactor(rust): 全面优化 v2.3.2 — 26 项改进（Bug/性能/UX/代码质量）
后端 Bug 修复：
- delete_user 事务性删除，避免崩溃产生孤儿数据
- login 去掉重复 find_user_by_username 查询
- duration_ms 由 i32 改为 i64，消除静默溢出风险
- cleanup_runs 校验 keepLatest 范围 1~10000
- ensure_setting_columns 从每次 get_settings 移到启动时

后端性能优化：
- create_user/account/run 改用 INSERT...RETURNING *
- scheduler 批量 COUNT 替代逐账户查询
- fetch_account_balance 路由复用 runner 函数
- 添加 idx_checkin_account_enabled 索引

后端代码质量：
- 错误消息全部统一为中文
- read_number/read_error_message/is_already_checked_message 提取到 providers/mod.rs
- base_url/site_type/name 输入校验
- admin.rs 移除多余 Result 别名
- Schema 添加 CHECK 约束（role/siteType/authType/status/triggeredBy）
- dotenv→dotenvy, lazy_static→std::sync::LazyLock
- 清理死代码：decrypt_secret、list_runs_by_user
- Cargo.toml 版本号对齐 CHANGELOG

前端修复：
- api.ts: 401 响应自动清除 token 刷新页面
- App.vue: login/fetchCurrentUser/logout 统一使用 request()
- AdminUserPanel/SettingsPanel/CheckinRunsPanel 添加 loading 状态
- AccountPanel 批量签到添加确认弹窗
- 模态框支持 Escape 键和点击遮罩关闭
- 响应式设计：navbar + SettingsPanel 支持 768px 断点
- 提取 types.ts + useUsers composable 消除重复代码
- 主色统一为 #0070f3，消除所有 !important
- 删除未使用资源文件

+++ src/models.rs
@@ -85,7 +85,7 @@ pub struct CheckinRun {
     pub message: Option<String>,
     #[serde(rename = "durationMs")]
     #[sqlx(rename = "durationMs")]
-    pub duration_ms: Option<i32>,
+    pub duration_ms: Option<i64>,
     #[serde(rename = "triggeredBy")]
     #[sqlx(rename = "triggeredBy")]
     pub triggered_by: String,
